LayoutTests/editing/undomanager/undomanager-undo-redo.html

<!DOCTYPE html>
<html>
<head>
<title>UndoManager has undo() and redo() that work</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
This tests UndoManager's undo(), redo(), length, and position.

<div id="edit" contentEditable="true"></div>
<script>

test(function() {
    assert_idl_attribute(document.undoManager, "undo");
    assert_idl_attribute(document.undoManager, "redo");
    assert_own_property(document.undoManager, "length");
    assert_own_property(document.undoManager, "position");
}, "undoManager has properties undo, redo, length, and position.");

test(function() {
    assert_true(document.undoManager.length == 0);
    assert_true(document.undoManager.position == 0);
}, "undoManager has correct initial values for length (0) and position (0).");

test(function() {
    var edit = document.getElementById("edit");
    window.getSelection().setPosition(edit, 0);
    document.execCommand("InsertHTML", false, "test");
    
    assert_true(document.undoManager.length == 1);
    assert_true(document.undoManager.position == 0);
}, "After inserting text 'test' in a DIV, undoManager's length is 1 and position is 0.");

test(function() {
    var range = document.createRange();
    range.selectNodeContents(edit);
    window.getSelection().addRange(range);
    document.execCommand("Bold", false, null);
    
    assert_true(document.undoManager.length == 2);
    assert_true(document.undoManager.position == 0);
    assert_true(edit.innerHTML == "<b>test</b>");
}, "After bolding the text, undoManager's length is 2 and position is 0, and the result is a bold 'test'.");
    
test(function() {
    document.undoManager.undo();
    
    assert_true(document.undoManager.length == 2);
    assert_true(document.undoManager.position == 1);
    assert_true(edit.innerHTML == "test");
}, "After undoing, undoManager's length is 2 and position is 1, and the result is a plain 'test'.");
    
test(function() {
    document.undoManager.undo();
    
    assert_true(document.undoManager.length == 2);
    assert_true(document.undoManager.position == 2);
    assert_true(edit.innerHTML == "");
}, "After undoing, undoManager's length is 2 and position is 2, and the result is empty.");
    
test(function() {
    document.undoManager.redo();
    
    assert_true(document.undoManager.length == 2);
    assert_true(document.undoManager.position == 1);
    assert_true(edit.innerHTML == "test");
}, "After redoing, undoManager's length is 2 and position is 1, and the result is a plain 'test'.");
    
test(function() {
    document.undoManager.redo();
    
    assert_true(document.undoManager.length == 2);
    assert_true(document.undoManager.position == 0);
    assert_true(edit.innerHTML == "<b>test</b>");
}, "After redoing, undoManager's length is 2 and position is 0, and the result is a bold 'test'.");

edit.style.display = 'none';

</script>
</body>
</html>
